add f64 support
